feat: Reorganize test suite
* move `helpers` out of src and into __tests__
* move sqs mocks out of `microservice.spec.ts` and into abstract methods in helpers
* ignore the helpers file when running tests
