+++ jest.config.js
@@ -7,4 +7,5 @@ module.exports = {
     // via https://github.com/axios/axios/issues/5101#issuecomment-1276572468
     '^axios$': require.resolve('axios'),
   },
+  modulePathIgnorePatterns: ['helpers.ts']
 };
